kdd2026/requirements.txt
# ==========================================
# Metalens FNO Project - Requirements
# ==========================================

# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Scientific Computing
numpy>=1.19.0
scipy>=1.7.0

# Configuration
pyyaml>=5.4.0

# Visualization
matplotlib>=3.4.0

# Progress Bar (optional)
tqdm>=4.60.0

# ==========================================
# Installation
# ==========================================
# pip install -r requirements.txt
#
# For CUDA support:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# ==========================================
